requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
flake8==6.0.0
flake8-django==1.2
flake8-import-order==0.18.2
install==1.3.5
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
sqlparse==0.4.4
